+++ packages/core/src/core.ts
@@ -6,6 +6,7 @@ import {
   AnonAadhaarProof,
   AnonAadhaarArgs,
   ArtifactsOrigin,
+  ProverState,
 } from './types'
 
 import { v4 as uuidv4 } from 'uuid'
@@ -38,7 +39,10 @@ export async function init(args: InitArgs): Promise<void> {
   initArgs = args
 }
 
-export async function prove(args: AnonAadhaarArgs): Promise<AnonAadhaarCore> {
+export async function prove(
+  args: AnonAadhaarArgs,
+  updateState?: (state: ProverState) => void
+): Promise<AnonAadhaarCore> {
   if (!initArgs) {
     throw new Error(
       'cannot make Anon Aadhaar proof: init has not been called yet'
@@ -66,7 +70,7 @@ export async function prove(args: AnonAadhaarArgs): Promise<AnonAadhaarCore> {
     initArgs.artifactsOrigin
   )
 
-  const anonAadhaarProof = await prover.proving(args)
+  const anonAadhaarProof = await prover.proving(args, updateState)
 
   return new AnonAadhaarCore(id, anonAadhaarClaim, anonAadhaarProof)
 }